Extract dataset workflow cluster into engine/src/depth/datasetWorkflow.ts
Moves triggerDatasetCalibrationLoading, triggerDatasetImageLoading, and
handleDatasetTexture out of main.ts. Deletes loadDatasetTexture, which
had zero call sites anywhere in the repo.
